requirements.txt
pydantic>=2.0,<3.0
openai>=1.0,<2.0
openenv-core>=0.1.0
python-dotenv>=0.19
pyyaml>=6.0
